refactor: Change user books state to be managed by App.js
User books list state were being managed by the BooksList component,
wich had the responsibility to render the books and call the BooksApi
on behalf of the state changes. This led to code repetition on the
BookSearch component, since it had to also manage book shelf updates
and by requirement it should know the state of the books in the main
page.

In addition, component names change and add a sortBooks helper
function.